Add a major mode for viewing timelines
The major mode is just derived from fundamental mode and does nothing
but it at least provides a keymap and a mode hook.

The default keymap sets C-c C-s to edit your status and C-c C-r to
reply to the status at point. The old reply button function is split
into a new interactive function called twitter-reply which uses the
text properties at point to set up a reply. The old callback function
now just calls this with the position of the reply button.