Display the tweets according to a customizable format string.
The format string can be customized using the twitter-status-format
variable. It takes a string similar to the 'format' function but it
has special commands to extract fields from tweets. Currently only a
limited number of commands are supported. Text properties are copied
from the format string. The default tries to mimic the original
behaviour as closely as possible.