Implemented sending direct messages.
Direct messages buffer is opened with C-c C-d. C-c C-r replies to a direct message (this is not a reply per se, but the recipients name is automatically added) and C-c C-p posts a new direct message (you will be queried for the username of the recipient).

C-c <down> and C-c <up> now works in direct messages buffer as well.

The direct messages buffer now contains both sent and recieved direct messages.